Moved ohlc series into a subfolder.
