+++ js/Series/OHLC/OHLCSeries.js
@@ -21,11 +21,11 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
-import ColumnSeries from './Column/ColumnSeries.js';
+import BaseSeries from '../../Core/Series/Series.js';
+import ColumnSeries from '../Column/ColumnSeries.js';
 var columnProto = ColumnSeries.prototype;
-import Point from '../Core/Series/Point.js';
-import U from '../Core/Utilities.js';
+import Point from '../../Core/Series/Point.js';
+import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
 /* *
  *
